Replaced all commands class to specific classes
