+++ src/main/java/io/github/lofrol/UselessClan/ClanCommands/ClanCommand.java
@@ -1,16 +1,14 @@
 package io.github.lofrol.UselessClan.ClanCommands;
 
 import io.github.lofrol.UselessClan.ClanCommands.Commands.CommandBase;
-import io.github.lofrol.UselessClan.ClanCommands.Commands.UserCommands.BaseClanCommands;
+import io.github.lofrol.UselessClan.ClanCommands.Commands.CommandsManager.BaseClanCommands;
 import io.github.lofrol.UselessClan.Utils.ChatSender;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
