+++ src/main/java/io/github/lofrol/UselessClan/ClanCommands/Commands/DefaultCommands/deleteUserCommand.java
@@ -1,0 +1,39 @@
+package io.github.lofrol.UselessClan.ClanCommands.Commands.DefaultCommands;
+
+import io.github.lofrol.UselessClan.ClanCommands.Commands.PlayerCommandBase;
+import io.github.lofrol.UselessClan.ClanObjects.Clan;
+import io.github.lofrol.UselessClan.ClanObjects.EClanRole;
+import io.github.lofrol.UselessClan.UselessClan;
+import io.github.lofrol.UselessClan.Utils.ChatSender;
+import org.bukkit.entity.Player;
+import org.jetbrains.annotations.NotNull;
+
+public class deleteUserCommand extends PlayerCommandBase {
+    @Override
+    public @NotNull String commandDescription() {
+        return "&a/Clan delete&b - to delete your clan";
+    }
+
+    @Override
+    public boolean havePermission(Player tempPlayer, Clan senderClan, EClanRole senderRole) {
+        return (senderRole == EClanRole.LEADER);
+    }
+
+    @Override
+    public boolean executeCommand(Player tempPlayer, Clan senderClan, String[] args) {
+        if (senderClan == null) {
+            ChatSender.MessageTo(tempPlayer, "UselessClan", "&cYou haven't Clan!");
+            return false;
+        }
+        EClanRole SenderRole = senderClan.getMemberRole(tempPlayer.getName());
+
+        if (SenderRole != EClanRole.LEADER) {
+            ChatSender.MessageTo(tempPlayer, "UselessClan", "&cYou rank is too low to do that!");
+            return false;
+        }
+
+        UselessClan.getMainManager().DeleteClan(senderClan);
+        ChatSender.MessageTo(tempPlayer, "UselessClan", "&aYou successfully delete your clan!");
+        return true;
+    }
+}
