+++ src/main/java/io/github/lofrol/UselessClan/ClanCommands/Commands/DefaultCommands/homeUserCommand.java
@@ -1,0 +1,45 @@
+package io.github.lofrol.UselessClan.ClanCommands.Commands.DefaultCommands;
+
+import io.github.lofrol.UselessClan.ClanCommands.Commands.PlayerCommandBase;
+import io.github.lofrol.UselessClan.ClanObjects.Clan;
+import io.github.lofrol.UselessClan.ClanObjects.EClanRole;
+import io.github.lofrol.UselessClan.Utils.ChatSender;
+import org.bukkit.Location;
+import org.bukkit.entity.Player;
+import org.jetbrains.annotations.NotNull;
+
+public class homeUserCommand extends PlayerCommandBase {
+
+    @Override
+    public @NotNull String commandDescription() {
+        return "&a/Clan home&b - to teleport to home of your clan";
+    }
+
+    @Override
+    public boolean havePermission(Player tempPlayer, Clan senderClan, EClanRole senderRole) {
+        return true;
+    }
+
+    @Override
+    public boolean executeCommand(Player tempPlayer, Clan senderClan, String[] args) {
+        if (senderClan == null) {
+            ChatSender.MessageTo(tempPlayer, "UselessClan", "&cYou havent Clan!");
+            return false;
+        }
+
+        Location tempHome = senderClan.getHomeClan();
+        if (tempHome == null) {
+            ChatSender.MessageTo(tempPlayer, "UselessClan", "&cYour clan doesnt have home!");
+            return false;
+        }
+
+        if (!tempPlayer.getWorld().getName().equals("world")) {
+            ChatSender.MessageTo(tempPlayer, "UselessClan", "&cYou cant teleport clan home from this world!");
+            return false;
+        }
+
+        tempPlayer.teleport(tempHome);
+        ChatSender.MessageTo(tempPlayer, "UselessClan", "&aYou teleported to clan home!");
+        return true;
+    }
+}
